Summarize a Compound Poisson-Normal (CPN) Model Fit

Produces a summary for a fitted cpn model object, including parameter estimates, standard errors, z-values, p-values, deviance residual summaries, and model diagnostics.

Usage

# S3 method for class 'cpn'
summary(object, ...)

Arguments

object

An object of class "cpn", see cpn().

...

Additional arguments (currently ignored).

Value

An object of class summary.cpn, which is a list containing:

call

The matched call that generated the model.

summary_table

A data frame with parameter estimates, standard errors, z-values, and p-values.

deviance_summary

A five-number summary (Min, 1Q, Median, 3Q, Max) of deviance residuals.

mu

Estimated value of the Normal component mean.

sigma

Estimated value of the Normal component standard deviation.

null_deviance

Null deviance of the model.

residual_deviance

Residual deviance of the model.

df_null

Degrees of freedom for the null model.

df_residual

Degrees of freedom for the residual model.

aic

Akaike Information Criterion for model comparison.

See also

Examples

set.seed(123)
n <- 100
x <- rnorm(n)
lambda <- exp(0.5 + 0.2 * x)
N <- rpois(n, lambda) # nolint
y <- ifelse(N == 0, 0, rnorm(n, mean = N * 3, sd = sqrt(N) * 2))
fit <- cpn(y ~ x)
summary(fit)
#> Call:
#> cpn(formula = y ~ x)
#> 
#> Deviance Residuals:
#>    Min     1Q Median     3Q    Max 
#> -2.741 -2.140 -1.681  2.407  3.149 
#> 
#> Coefficients:
#>             Estimate Std.Error z.value      Pr.z    
#> (Intercept)  0.49166   0.12165  4.0414 5.313e-05 ***
#> x            0.20931   0.09367  2.2345   0.02545 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Estimated mu parameter: 2.9067
#> Estimated sigma parameter: 1.5520
#> 
#> Null deviance: 534.45 on 99 degrees of freedom
#> Residual deviance: 529.60 on 96 degrees of freedom
#> AIC: 537.60